Problem 11 - 14

+++ Problem1_Multiples35.cpp
@@ -8,10 +8,10 @@ bool IsMultiple(int i)
     return (i % 3 == 0) || (i % 5 == 0);
 }
 
-int Problem_1_Multiples_3_5()
+long long Problem_1_Multiples_3_5()
 {
-    std::vector<int> range = generate_range(0, 1000);
-    std::vector<int> numbers;
+    std::vector<long long> range = generate_range(0LL, 1000LL);
+    std::vector<long long> numbers;
     auto isMultiple = [&numbers] (int i)
     {
       if (IsMultiple(i))
